Fixed memory error
Changed readKeyFromFile to use a temporary array to read the characters before casting it and putting it in the correct key array. For some reason this was sometimes causing major issues before and tbh I haven't tested enough yet to know if it's actually fixed or I just got lucky and it started working when I made this change.